pyproject.toml
[project]
name = "codeagent"
version = "0.1.0"
description = "AI-powered coding agent"
requires-python = ">=3.10"
dependencies = [
    "google-genai",
    "rich>=13.0.0",
]

[project.scripts]
codeagent = "codeagent.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"